Minor change to level 1 and 5 + add global score script and show Finish game message after 5 levels

+++ Assets/Scripts/ShootLogicV3.cs
@@ -5,7 +5,7 @@ public class ShootLogicV3 : MonoBehaviour {
 
     public float maxDrag = 2.0f;
     public float lineWidth = 0.30f;
-    public int acceleration = 2;
+    public float acceleration = 2.0f;
 
     private float maxDragSqr;
